requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
requests>=2.31.0
trafilatura>=1.6.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

